pyproject.toml
[project]
name = "pdfbeta"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "bcrypt==4.0.1",
    "boto3>=1.42.59",
    "fastapi>=0.128.8",
    "passlib[bcrypt]>=1.7.4",
    "pdfplumber>=0.11.9",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.47",
    "uvicorn>=0.40.0",
]
